requirements.txt
pandas==2.1.1
selenium==4.18.1
python-dotenv==1.0.1
smtplib